+++ app/Models/Observers/ComputerObserver.php
@@ -13,18 +13,20 @@ class ComputerObserver extends Observer
      */
     public function deleting(Computer $computer)
     {
-        $computer->access()->delete();
+        if (!$computer->deleted_at) {
+            $computer->access()->delete();
 
-        $disks = $computer->disks()->get();
+            $disks = $computer->disks()->get();
 
-        foreach ($disks as $disk) {
-            $disk->delete();
-        }
+            foreach ($disks as $disk) {
+                $disk->delete();
+            }
 
-        $statuses = $computer->statuses()->get();
+            $statuses = $computer->statuses()->get();
 
-        foreach ($statuses as $status) {
-            $status->delete();
+            foreach ($statuses as $status) {
+                $status->delete();
+            }
         }
     }
 }
